requirements.txt
Flask==2.0.1
numpy==1.23.5
pandas==1.5.3
gymnasium==0.28.1
stable-baselines3==2.1.0
matplotlib==3.7.1
seaborn==0.12.2
scipy==1.10.1
torch==2.0.1
requests==2.31.0
scikit-learn==1.2.2
beautifulsoup4==4.12.2
plotly==5.18.0
joblib==1.3.2
pytest==7.4.0
APScheduler==3.10.1
Werkzeug==2.3.7
Jinja2==3.1.2
gunicorn==21.2.0
python-dotenv==1.0.0
Flask-Cors==4.0.0
markupsafe==2.1.3 